// src/api/model/channelInputDto.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ChannelUserDto } from "./channelUserDto";
import type { ChannelCategoryDto } from "./channelCategoryDto";
import type { ChannelInputDtoTechStack } from "./channelInputDtoTechStack";
import type { ChannelInputDtoSolution } from "./channelInputDtoSolution";

export interface ChannelInputDto {
  id?: string;
  channelId?: number;
  link?: string;
  channelUser?: ChannelUserDto;
  category?: ChannelCategoryDto;
  techStack?: ChannelInputDtoTechStack;
  participantsCount?: number;
  ic?: number;
  irr?: number;
  connected?: number;
  price?: number;
  personPrice?: number;
  comment?: string;
  utmUrl?: string;
  shortUrl?: string;
  startedDate?: string;
  registration?: number;
  registrationPrice?: number;
  solution?: ChannelInputDtoSolution;
  paymentMethod?: string;
  redLabel?: boolean;
  blackLabel?: boolean;
  greenLabel?: boolean;
  estimated?: number;
}